### 2023-02-21 #5179 @editorialbot
[PRE REVIEW]: Chitin Builder: a VMD tool for the generation of structures of chitin molecular crystals for atomistic simulations
**Submitting author:** @jfaraudo (<a href="http://orcid.org/0000-0002-6315-4993">JORDI FARAUDO</a>)
**Repository:** https://github.com/soft-matter-theory-at-icmab-csic/chitin_builder
**Branch with paper.md** (empty if default branch):
**Version:** 1.0
