### 2023-03-06 #5179 @majensen
Hi @symmy596 - would you have time and willingness to review [this work](#5179) for [JOSS](https://joss.theoj.org)? I could use your expertise - thanks for considering.
-- @majensen
