### 2023-05-23 #5179 @majensen
@dandavies99 - would you have time to review [this submission](#5179) for [JOSS](https://joss.theoj.org)? I appreciate your consideration - thanks. --@majensen
