### 2023-06-11 #5179 @majensen
Hey @TomKellyGenetics - would you have time and interest in reviewing [this work](#5179) for [JOSS](https://joss.theoj.org)? Thanks for considering- --@majensen
