### 2023-08-20 #5179 @majensen
@WangKehan573 - would you be able to review [this work](#5179) for the journal [JOSS](https://joss.theoj.org)? I would welcome your expertise with the VMD framework here. Thanks --@majensen
