### 2023-08-21 #5179 @arfon
The top 5 most similar JOSS papers are:

[ViMMS 2.0: A framework to develop, test and optimise fragmentation strategies in LC-MS metabolomics](https://github.com/openjournals/joss-reviews/issues/3990)
Reviewers: `@jspaezp, @MKoesters`
Similarity score: 0.795775191754251

[MTEX2Gmsh: a tool for generating 2D meshes from EBSD data](https://github.com/openjournals/joss-reviews/issues/2094)
Reviewers: `@streeve, @ralfHielscher`
Similarity score: 0.7901755944085497

[GEMMI: A library for structural biology](https://github.com/openjournals/joss-reviews/issues/4200)
Reviewers: `@jamesrhester, @dominiquesydow`
Similarity score: 0.78667981441994

[FHI-vibes: Ab Initio Vibrational Simulations](https://github.com/openjournals/joss-reviews/issues/2671)
Reviewers: `@keipertk, @ajjackson`
Similarity score: 0.7852457526328093

[Chemiscope: interactive structure-property explorer for materials and molecules](https://github.com/openjournals/joss-reviews/issues/2117)
Reviewers: `@kblin, @mkhorton`
Similarity score: 0.7840161893561871
