### 2023-10-15 #5771 @amoeba
Hi @majensen, while reviewing the "Substantial scholarly effort" checklist item, I find a could use some editorial help. Using the [corresponding section in the Review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#substantial-scholarly-effort) as a guide, I think this submission is borderline due to the size and apparent age of the codebase. I don't see that this was discussed in pre review but please let me know if I missed that

It seems to me the most important element of this checklist item is whether the work is of broad interest and likely to be cited which isn't something I can assess with my background. However, in the paper, the authors speak to this at the end of their _Introduction and Statement of Need_ section and to me it sounds like they theorize this tool would be quite useful to others in their field. With that, I'm inclined to check "Substantial scholarly effort" so long as you agree.
