### 2023-10-15 soft-matter-theory-at-icmab-csic/chitin_builder#3 @amoeba
JOSS Review
Hi @jfaraudo, I've started [my review](https://github.com/openjournals/joss-reviews/issues/5771#issuecomment-1763121112) and will file issues here. I'm working on the review now so I'm going to edit this comment as I go.

The following review checklist items require changes:

  - Requested Change: Please use the uppercase `LICENSE` for the license file's name
- [x] **Paper:** The paper has far too much content at this point and should be shortened considerably. Please review https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain. I think you could reduce your _Introduction and Statement of Need_ section into just a _Statement of Need_ section into three paragraphs, mostly containing the text at the end of the existing section. Remove _Methods_, _Program description, use and features_, _Examples_, _Compatibility and distribution_.

Non-review-checklist items:

- [ ] Update Zenodo citation to match auto-generated citation
- [x] Try to use the name "Chitin Builder" everywhere, updating any alternative usages
- [x] Had an issue opening the documentation from the Chitin Builder GUI. Filed https://github.com/soft-matter-theory-at-icmab-csic/chitin_builder/issues/4.
