### 2023-10-15 soft-matter-theory-at-icmab-csic/chitin_builder#3 @amoeba
Hi @jfaraudo, I ran into this error when I opened Chitin Builder and clicked the "Generate Chitin Structure" button. I'm having this issue both on macOS and Windows:

```
invalid command name "::namespace inscope :: {::chitin::replicate "$::chitin::xdim" "$::chitin::ydim" "$::chitin::zdim" "$::chitin::crystal" "$::chitin::perio"}"
invalid command name "::namespace inscope :: {::chitin::replicate "$::chitin::xdim" "$::chitin::ydim" "$::chitin::zdim" "$::chitin::crystal" "$::chitin::perio"}"
    while executing
"[namespace code {::chitin::replicate "$::chitin::xdim" "$::chitin::ydim" "$::chitin::zdim" "$::chitin::crystal" "$::chitin::perio"}]"
    invoked from within
".chitin.but46 invoke"
    ("uplevel" body line 1)
    invoked from within
"uplevel #0 [list $w invoke]"
    (procedure "tk::ButtonUp" line 23)
    invoked from within
"tk::ButtonUp .chitin.but46"
    (command bound to event)
```

![Screenshot 2023-10-15 at 3 56 45 PM](https://github.com/soft-matter-theory-at-icmab-csic/chitin_builder/assets/563/d71178fe-107c-4ca1-ab5b-ef3ecfda6eb9)

any ideas?
