### 2023-10-16 soft-matter-theory-at-icmab-csic/chitin_builder#3 @jfaraudo
hello,
I cannot reproduce this error. I tried with my Mac and Ubuntu.
Which OS and version are you using?
best
jordi

From: "Bryce Mecum" ***@***.***>
To: "soft-matter-theory-at-icmab-csic/chitin_builder" ***@***.***>
Cc: "Jordi Faraudo" ***@***.***>, "Mention" ***@***.***>
Sent: Monday, 16 October, 2023 01:58:29
Subject: Re: [soft-matter-theory-at-icmab-csic/chitin_builder] JOSS Review (Issue #3)

Hi [ https://urldefense.com/v3/__https://github.com/jfaraudo__;!!D9dNQwwGXtA!UTY2cbSfEbQa6NbmEjClCx7M9jCM-_JRipFhTOCNYmSctIn__At2UNUs8Q9PEpGwZHCc4Vd7DIGU2kWkHrq5Vw1Uxg$ | @jfaraudo ] , I ran into this error when I opened Chitin Builder and clicked the "Generate Chitin Structure" button:
invalid command name "::namespace inscope :: {::chitin::replicate "$::chitin::xdim" "$::chitin::ydim" "$::chitin::zdim" "$::chitin::crystal" "$::chitin::perio"}"
invalid command name "::namespace inscope :: {::chitin::replicate "$::chitin::xdim" "$::chitin::ydim" "$::chitin::zdim" "$::chitin::crystal" "$::chitin::perio"}"
    while executing
"[namespace code {::chitin::replicate "$::chitin::xdim" "$::chitin::ydim" "$::chitin::zdim" "$::chitin::crystal" "$::chitin::perio"}]"
    invoked from within
".chitin.but46 invoke"
    ("uplevel" body line 1)
    invoked from within
"uplevel #0 [list $w invoke]"
    (procedure "tk::ButtonUp" line 23)
    invoked from within
"tk::ButtonUp .chitin.but46"
    (command bound to event)

[ https://urldefense.com/v3/__https://user-images.githubusercontent.com/563/275362184-d71178fe-107c-4ca1-ab5b-ef3ecfda6eb9.png__;!!D9dNQwwGXtA!UTY2cbSfEbQa6NbmEjClCx7M9jCM-_JRipFhTOCNYmSctIn__At2UNUs8Q9PEpGwZHCc4Vd7DIGU2kWkHrpnRrRNLw$ ]

any ideas?
