### 2023-10-16 soft-matter-theory-at-icmab-csic/chitin_builder#3 @jfaraudo
Hello @amoeba , sorry to hear this.
We had some users using WIndows that reported problems after an update 1 year ago and we have been unable to solve it , the only workout that we had was to ask users to move to Ubuntu or Mac.
Up to now we had no problems reported for Mac users, so I am worried to see that you had problems in a Mac. Please, could you
1) quote the exact OS and version that you are running?
2) It looks like a permisions issue in creating the crystal structure. Maybe running VMD as superuser may help.
thanks!
jordi
PS: Also, I am curios to know what happens if you accept (OK or skip). Are the pdb and psf files of the structure generated? and displayed? This may help to locate the issue.
Thank you!!
