### 2023-10-16 soft-matter-theory-at-icmab-csic/chitin_builder#3 @jfaraudo
Hello again @amoeba ,
I have been able to reproduce your error on a macOS Monterey 12.4 . For some reason, in our older Mac this error does not appear. I also checked that the same error appears in WIndows 11. The code works fine in Ubuntu 18 and 20.
We have found that the error appears when the code is working with a temporary file that is automatically removed before the final structure is copied to the destination file selected by the user. We believe that this can be fixed easily by a simple modification in the tcl code when dealing with the temporary files.
We will try to deal with this problem as soon as possible and provide a modified tcl file that hopefully will solve this issue for Mac and windows users.

I assume that you do not have a machine with ubuntu handy, so I think the best option for you to continue your evaluation should be to wait until we post the correction .

Thanks again for your time testing the "chitin builder".
