### 2023-10-16 soft-matter-theory-at-icmab-csic/chitin_builder#3 @amoeba
Hi @jfaraudo, thanks for tracking down the issue. I can continue my review using a Linux VM for now and can help verify your fix on macOS and Windows when it's ready. I should have time to continue that this weekend. Let me know if you have any questions about my request for edits on the paper in the mean time.
