### 2023-10-17 soft-matter-theory-at-icmab-csic/chitin_builder#3 @jfaraudo
Hi @amoeba ,
Today I was going to start the fix for Mac but before that, I changed the permisions of the files and now it works. I am attaching a picture with the permissions that I have in my Mac. My MacOS is Monterey.
