### 2023-11-02 #5771 @majensen
Sorry for the long delay @amoeba  and all. I am discussing this with one of the AEICs. @jfaraudo - can you comment here on your view of the effort put into this software, and on its usefulness to the community? I think it is a specialized community, and that is ok - how do you see this work contributing to the science?
