### 2023-11-08 #5771 @jfaraudo
@majensen Yes, sure, we can comment on this here.
Yes, the code is useful for a very specialized community. Our target are the VMD users that develop simulations of polymeric materials based on natural polymers. To make a comparison, it is an scope similar to that of the "cellulose builder" app, that generates structures cellulosic materials, but in our case for chitin based materials. The relevance of this is based on the fact that cellulose and chitin are the most abundant polymers in Earth and are both of natural origin and therefore interesting as "green" materials. They have also very interesting properties. Prediction by computer simulation of their properties and behaviour is relatively under-studied (compared with other less abundant materials or compared with other human-designed materials) due to practical difficulties in generating appropiate structures suitable for the simulations. The objective of this code is to solve this problem, providing interested researchers with a tool that is able to generate the necessary files to investigate by computer simulations the properties of materials based on different crsytall structures of chitin polymer. We hope that this app will help in the advancement of the understanding of chitin and rational design materials based on this natural polymer by computer simulation.

> @jfaraudo - can you comment here on your view of the effort put into this software, and on its usefulness to the community? I think it is a specialized community, and that is ok - how do you see this work contributing to the science?
