### 2023-11-16 #5771 @majensen
Thanks @jfaraudo - I have also spent some time getting a deeper understanding of the VMD framework, which it is very current use, with new plugins written for it frequently up to the present time (even though it was published first around 1996). @amoeba I would call it the "Emacs of Molecular Modeling". I have not heard any disagreement from the Associate Editor in Chief to this argument. So let's proceed with "Substantial scholarly effort".
