### 2023-11-16 #5771 @amoeba
>  So let's proceed with "Substantial scholarly effort".

Thanks for taking a look, @majensen. Works for me.
