### 2023-11-23 #5771 @jfaraudo
Sorry for the delay in responding. I think apropiate referees could be developers of other VMD plugins. One example is Dr Toni Giorgino @tonigi [ https://github.com/tonigi | https://github.com/tonigi ]
Let me think about others.
jordi

From: "Mark Jensen" ***@***.***>
To: "openjournals/joss-reviews" ***@***.***>
Cc: "Jordi Faraudo" ***@***.***>, "Mention" ***@***.***>
Sent: Thursday, 16 November, 2023 23:15:28
Subject: Re: [openjournals/joss-reviews] [REVIEW]: Chitin Builder: a VMD tool for the generation of structures of chitin molecular crystals for atomistic simulations (Issue #5771)

[ https://urldefense.com/v3/__https://github.com/jfaraudo__;!!D9dNQwwGXtA!SoVYcjZdOWxeXeN0LHD9qH4oQ6jNIU5hcjJUbM0yWsXZW_b9OuhoxRVV66y2u82gJIOXk1ZcVLfopsmGnCnIN695DQ$ | @jfaraudo ] -- unfortunately one of our reviewers had to decline. I am sorry about the very long delay. Would you have any suggestions for reviewers with the right expertise who could fairly review your work? Much appreciated!
