### 2023-11-24 soft-matter-theory-at-icmab-csic/chitin_builder#3 @tonigi
Here is my review and comments. (Let me know if it is not the right place).

* The plugin is indeed a relatively short piece of software. However, it is not uncommon for VMD plugins to perform useful tasks in a few lines of code. In my opinion the sheer amount of code should not be a factor in the decision for this kind of software.
* On the contrary, the ability to conveniently build structured materials for simulations is a common need for simulations. Of course every community will have their material of choice, so perhaps not many people are working on chitin, but it is very good to have a builder for it.
* I did not encounter the mentioned bugs with MacOS, except some tk-related quirkiness which is kind of expected and outside of the author's control.
* The paper reads nicely.

On the other hand, the software *implementation* could benefit from some additional polish and a software-savvy eye.
* As mentioned, the documentation points to an obsolete page.
* It's slow in building, even a 5x5x5 supercell. 11x11x1 takes ages on a fast machine. That's weird given that systems are < 30k atoms.
* The software structure is unnecessarily involved: there is no separation between gui and code (e.g. the "main" calculation is all done in an inline code block within the "Generate" button)
* Most important of all, there is no way to invoke the generator from the command line or other software (dual GUI and command line use, as is common for VMD plugins)
* Looking at the `replicate` function, which seems the main entry point, some arguments are passed by argument, some are taken from the environment.
* There are some "duplicate atom key" warnings upon second and following runs. This may suggest a skipped "resetpsf"
* Every generate requires navigating to the destination directory, every time starting from the system root. Also, clicking cancel does not abort the generation.
* The "crystal.log" file is often empty
