### 2023-11-24 #5771 @tonigi
I've put my review at https://github.com/soft-matter-theory-at-icmab-csic/chitin_builder/issues/3#issuecomment-1825364955 . The software may be improved but it mostly works. There is no "API" (yet there should be) so i'm undecided on the "Functionality documentation"; i'm checking it for the time being.
