### 2024-01-02 soft-matter-theory-at-icmab-csic/chitin_builder#3 @amoeba
Hi @jfaraudo, I just wanted to check in on the status of (1) your updates to the paper as requested above and (2) you posting a new software release for me to test on macOS. Let me know if I can help at all such as reviewing early edits or testing early releases.
