### 2024-01-03 soft-matter-theory-at-icmab-csic/chitin_builder#3 @jfaraudo
Hi @amoeba ! We just posted a new version at the master branch at chitin github repo GitHub https://github.com/soft-matter-theory-at-icmab-csic/chitin_builder  The new version works better than the previous one, it works fine in most of the tests I did in MAC but sometimes there is still an odd  behaviour (not always). Sometimes there is a write error when creating the output files which seems to depend on the OS version and the name of the output folder. We noticed that when this error appears, in can be easily skipped by going to the Extensions - Tk console menu in VMD, and in the console just do a cd to the desired folder (this should be done automatically but it seems that sometimes something is wrong with the appropiate environment global variable of VMD). If the error appears and one does this trick once, the error no longer appears (even closing VMD and opening it again). We are working on fixing this (hopefully it will be done soon), but if you can try with your computer your feedback will be appreciated.
