### 2024-01-03 62a635a
Update README.md
Updated description including troubleshooting
