### 2024-01-04 5f302ae
Update README.md
Added a warning for MAC OS users.
