### 2024-01-04 soft-matter-theory-at-icmab-csic/chitin_builder#3 @jfaraudo
Hi @amoeba ! We just posted our final version at the master branch at chitin github repo GitHub https://github.com/soft-matter-theory-at-icmab-csic/chitin_builder
Now it works without problems in WIndows (tested 10 and 11) . In the case of MAC, depending on the OS version , the Tk console of VMD must be open during the chitin builder execution. This simply means selecting the option Tk console in the VMD Extensions menu. We added an explanation about this in the README.
Now we will work on the paper changes.
