### 2024-01-04 #5771 @jfaraudo
Hello @tonigi @amoeba @majensen ,
We have updated the code following the recommendations and also solved some issues for Windows and MAC users (in the case of MAC users for some versions it is necessary to do an additional action before running the code due to a VMD bug, as explained in the README). We are now going to update and shorten the paper follwong the recommendations by @amoeba . We will notify as soon as it is ready (we preferred to focus first on the code issues). Some specific comments/responses are also given below.
