### 2024-01-04 #5771 @jfaraudo
@tonigi Concerning the speed of the code , we were not able to reproduce the slow speed reported in your example. We checked the speed in one of our desktop computers (Intel® Core™ i5-10400F CPU @ 2.90GHz × 12, NVIDIA Corporation TU116 [GeForce GTX 1650 SUPER]). It took 13 seconds to replicate a 11x11x11 alpha crystal (143,748 atoms). We think it is fairly fast, but of course the execution time is extremely device dependent and depends mostly on VMD code, such it is the one performing the calculations. We are not aware of any issue influencing the speed as the reviewer mentioned.
