### 2024-01-04 #5771 @jfaraudo
Concerning the code structure, we have tried to improve it following @tonigi @amoeba  comments.
1) We fixed the error with the the Documentation URL.
2) The `replicate` function was improved and moved some arguments to avoid conflicts between variables.
3) We have tried to reduce the warning messages from VMD. However, there are still some "duplicate atom key" warnings, which cannot be avoided, and should be ignored by the user. The reason is that they appear due to overlapping residues during the crystal building. These overlaps are then fixed with the patch command. The resetpsf command would not avoid these warnings since the “duplicate atom key” is just a result of how we replicate the crystal. In further updates we will look at a way to reduce these warning messages, but for the time being the simplest option is simply ignore it.
4) We fixed the requirement to navigate to the destination directory for every structure, every time starting from the system root.  Now the default directory location is ~ .
5) We fixed the cancel button, it now aborts the generation of the crystal structure.
6) There was an issue with the "crystal.log" file, which it is often empty. We fixed this issue, the log file was not properly closed.
