### 2024-01-04 #5771 @jfaraudo
The GUI and the main code are still all together in the same code, the script works invoking different procedures that can be called independently. We are thinking in reorganizing the code for future versions, but for now we are more focused on functionality.
Actually, besides that the GUI and the main code are together in the same script, you can invoke the main replication function “replicate” by just loading the package and setting the working directory. To run from the VMD Tk console you need to run the following commands:

package require chitin
set chitin::fname “/home/”
chitin::replicate n1 n2 n3 crys1 per1

Where “\home\” should be replaced with your working directory.
The arguments n1, n2 and n3 should be integers representing the number of replicas in x, y and z respectively.
The argument crys1 should be the string “Alpha” or “Beta” depending on the chitin crystal structure to replicate.
Finally the argument per1 should be the string “yes” or “no” depending if you want periodic bonds between the first and last residue.

We are aware of a bug that after using the Chitin Builder through the VMD command line, an error appears if afterwards you try to generate a crystal using the GUI. This bug can be worked around by just restarting VMD.
