### 2024-01-04 #5771 @jfaraudo
We think this is all concerning the code. We will now update/shorten the paper asap.
