### 2024-01-05 #5771 @majensen
Thanks for all this work @jfaraudo  - would you like to comment on @tonigi 's concern (https://github.com/openjournals/joss-reviews/issues/5771#issuecomment-1825367156) about the package having an API?
