### 2024-01-08 #5771 @jfaraudo
Thanks @majensen @tonigi . Sorry, I am not sure what exactly means an API in this context. But the commands from the tcl script can be called from outside the GUI, in particular from the Tk console, as we explained in the response (we will add further info on that in the revised paper). This means that the commands are also -in principle- callable from any other additional APP running in tcl in VMD. I think this should be enough, for the standard VMD simulation user/developer.
