### 2024-01-08 #5771 @tonigi
By API I meant (loosely speaking) instructions to invoke the builder from TCL (i.e., from other scripts, not the GUI). It appears to be addressed in this comment https://github.com/openjournals/joss-reviews/issues/5771#issuecomment-1877382167 which possibly I guess deserves a brief section in the documentation.
