### 2024-01-08 soft-matter-theory-at-icmab-csic/chitin_builder#3 @amoeba
Thanks @jfaraudo, I'll return to my review this weekend using the new version and will let you know how that goes.
