### 2024-01-09 #5771 @jfaraudo
@tonigi Fully agree. We will add instructions to the documentation and an example on how to use it.
