+++ examples/readme.md
@@ -1,0 +1,2 @@
+In this folder we provide the output files generated in the Examples section of the program manual in the documentation folder. 
+Please read the manual available here
