### 2024-01-09 22787bb
Update readme.md
Updated description of the user manual
