+++ examples/readme.md
@@ -1,2 +1,2 @@
 In this folder we provide the output files generated in the Examples section of the program manual in the documentation folder. 
-Please read the manual available here
+Please read the manual available here https://github.com/soft-matter-theory-at-icmab-csic/chitin_builder/tree/master/documentation 
