+++ structures/README.md
@@ -1,0 +1,7 @@
+This folder contains files employed to generate structures
+
+alpha-4residues.pdb: initial structure for alpha chitin.
+
+beta-4residues.pdb: initial structure for beta chitin.
+
+top_all36_carb.rtf: CHARMM36 carbohydrate topology file for generating psf files.
