+++ examples/README.md
@@ -1,2 +1,3 @@
 In this folder we provide the output files generated in the Examples section of the program manual in the documentation folder. 
+Each subfolder contains a screenshot of the input parameters and the output files.
 Please read the manual available here https://github.com/soft-matter-theory-at-icmab-csic/chitin_builder/tree/master/documentation 
