+++ README.md
@@ -14,10 +14,10 @@ For detailed documentation with the methodology please go to the "documentation"
 ----------------------------------------------------------------
 ## Citation
 
-*D.C. Malaspina and J. Faraudo (2019)
-[![DOI](https://zenodo.org/badge/195977540.svg)](https://zenodo.org/badge/latestdoi/195977540)
+Please cite the following JOSS article:
+[Submitted , see article folder]
 
------------------------------------------
+----------------------------------------------------------------
 ## Files
 
 * chitin-gui.tcl: main code file in tcl language.
@@ -117,6 +117,9 @@ Inside this tcl folder, create a new folder with the name chitin1.0
 
 
 ## Usage
+
+once installed, the chitin builder can be used from the GUI available as a new VMD option or from the VMD Tk command line (see the User Manual for details). 
+For new users, the simplest option is using the GUI interface which can be selected from VMD - Extensions menu.
 
 ![alternativetext](/examples/example-b-beta-2-2-4/beta-2-2-4.png)
 
