+++ examples/Tkconsole_example/README.md
@@ -1,0 +1,1 @@
+Screenshots showing typical input and output generating a chitin crystal from the command line.
