+++ README.md
@@ -120,6 +120,8 @@ Inside this tcl folder, create a new folder with the name chitin1.0
 
 Once installed, the chitin builder can be used from the GUI available as a new VMD option or from the VMD Tk command line (see the User Manual for details). 
 
+-------------------------------------------
+
 * **Usage from the Graphical User Interface**
   
 For new users, the simplest option is using the GUI interface which can be selected from VMD menu (select: Extensions - Modelling - Chitin Builder).
@@ -139,6 +141,8 @@ Data of the unit cell used for the generation of the new crystal would appear un
 4) **Generate chitin structure button:** will prompt a dialog box for the destination folder and will generate a PDB + PSF file named: "crystal-alpha-psf.pdb / psf or crystal-beta-psf.pdb / psf" with the crystal structure and topology file.
 
 The data for the new crystal cell would be print under the generate structure button and the same data will be print on the file "crystal.log".
+
+-------------------------------------------
 
 * **Usage from the Tk Console**
 
@@ -165,6 +169,9 @@ The argument per1 should be the string “yes” or “no” depending if you wa
 
 The following screenshots show typical input and output examples for generating a crystal using the Tk console (see the Manual for details).
 
+![alternativetext](/examples/Tkconsole_example/inputTk.png)
+
+![alternativetext](/examples/Tkconsole_example/outputTk.png)
 
 After using the Chitin Builder through the VMD command line, an error appears if afterwards you try to generate a crystal using the GUI. 
 If you wish to use the GUI, restart VMD.
