### 2024-01-09 #5771 @jfaraudo
@tonigi As suggested, we have added a brief comment in the README.md (with screenshots) and a brief section in the User Manual showing how to call the chitin builder from the VMD Tk console without using the GUI.
