+++ paper/paper.md
@@ -32,19 +32,12 @@ Crystal structures of any size with or without bonds in the periodic directions 
 The resulting structure is automatically saved in PDB and PSF format (used by well known simulation packages such as NAMD) and it can be easily converted to many other file formats using VMD build-in features.
 
 
-# Introduction and Statement of Need
-
-\label{sec:Introduction}
+# Statement of Need
 
 Chitin is a polysaccharide present in the exoskeleton and internal structure of many invertebrates like molluscs, crustaceans, insects, fungus, algae, and other related organisms [@Dutta2002CHITINAPPLICATIONS; @Zargar2015AApplications].
 It is so prevalent in nature that it constitutes the second most abundant polymerized form of carbon in Earth.
 From the point of view of material sciences, chitin is biorenewable, environmentally friendly, biocompatible and biodegradable material.
 It has applications as a chelating agent, water treatment additive, drug carrier, biodegradable pressure‐sensitive adhesive tape, wound‐healing agents and many others [@Zargar2015AApplications; @Shamshina2019AdvancesReview; @RaviKumar2000AApplications]. 
-
-Chemically, chitin is a polysaccharide composed of repeated units of (1-4) N-acetyl-D-glucosamine (2-acetamido-2-deoxy-$\beta$-D-glucopyranose). 
-As other polysaccharides such as cellulose, chitin is usually found in crystal structures with several inter-chain hydrogen bonds that produce high order structures. 
-There are three known crystal allomorphs [@Jang2004PhysicochemicalResources]: $\alpha$, $\beta$ and $\gamma$ chitin.
-The crystal structures of $\alpha$ and $\beta$ chitin (the most abundant allomorphs) are known [@Sikorski2009Revisit; @Nishiyama2011Xray] but the structure of the rare $\gamma$ chitin alomorph is still under debate [@Jang2004PhysicochemicalResources; @Kaya2017On-chitin; @Ramirez-Wong2016Sustainable-solvent-inducedFilms].
 
 \begin{figure}[ht]
 \includegraphics[width=0.9\columnwidth]{alpha-beta.png}
@@ -67,157 +60,18 @@ In this work we present a chitin builder tool implemented as a plugin of the Vis
 
 This plugin will greatly facilitate the process of generation of input files (coordinates, structures, topology) for atomistic simulations and we expect that it will fuel the use of these techniques in the study of these materials. 
 Future developments of these plugin will incorporate the generation of crystal structures of other polymeric crystals.
-
-# Methods
-
-\label{sec:Methods}
-\subsection{Crystal structures of Chitin}
-
-As we mentioned in the Introduction, there are three known chitin crystal allomorphs, $\alpha$, $\beta$ and $\gamma$ chitin.
-However, we will not consider $\gamma$ chitin since its structure is not well known and in fact it is sometimes considered as a mixture of $\alpha$ and $\beta$ chitin [@Jang2004PhysicochemicalResources; @Kaya2017On-chitin; @Ramirez-Wong2016Sustainable-solvent-inducedFilms].
-The crystal structures of $\alpha$ and $\beta$ chitin allomorphs are shown in Figure \ref{fgr:crystals}.
-In the case of $\alpha$ chitin, we show in Figure \ref{fgr:crystals}a the primitive cell as determined by from X-ray diffraction in Ref. [@Sikorski2009Revisit].
-This is the unit cell employed in our program for $\alpha$ chitin.
- Note that $\alpha$ chitin crystals have a characteristic antiparallel organization of chitin chains, as shown in Figure \ref{fgr:crystals}c. 
  
[…]
