### 2024-01-09 c35292d
Update paper.md
Modified paper taking into account referee comments
