+++ paper/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 # Summary
 Chitin is the second most abundant organic material in nature after cellulose and its study is now of great interest in the field of biocompatible and eco-friendly materials. 
 Computational studies of natural polymers such as Cellulose and Chitin are often hindered by the practical difficulties in generating structures suitable for the simulations.
-This has motivated the recent introduction of a cellulose builder tool that generates coordinate, structure and topology files for atomistic simulations of cellulose crystal polymorhs.
+This has motivated the recent introduction of a cellulose builder tool that generates coordinate, structure and topology files for atomistic simulations of cellulose crystal polymorphs.
 Here we present an analogous tool for Chitin, the Chitin Builder tool, a program that enhances the Visual Molecular Dynamics (VMD) environment with the ability to generate coordinate and structure files of chitin organic crystals.
 The program generates cartesian coordinates and atomic connectivity and structure files for crystalline structures of chitin polymorphs $\alpha$ and $\beta$.
 Crystal structures of any size with or without bonds in the periodic directions can be easily build.
