### 2024-01-10 38e2bdd
Update paper.md
A few typos corrected
