### 2024-01-10 #5771 @jfaraudo
@majensen We have corrected the typos indicated by @tonigi . Concerning the origin of the unit cells of the crystal structures , as mentioned before, they are not from PDB database so they do not have a PDB ID. They were .cif files provided in crystallographic publications and we converted them to pdb as described in the Method sectionof the user manual. We have added a line in the paper indicating this fact and pointing to the methods section in the User Manual for further technical details.
