### 2024-01-15 #5771 @amoeba
Hey @majensen, I'm re-starting my review now and will report back shortly. At the very least, a new release of the software needs to be made and the JOSS submission needs to be updated.
