### 2024-01-15 8c2f329
Perform an editing pass over paper.md
