+++ paper/paper.md
@@ -21,55 +21,55 @@ bibliography: paper.bib
 ---
 
 # Summary
-Chitin is the second most abundant organic material in nature after cellulose and its study is now of great interest in the field of biocompatible and eco-friendly materials. 
-Computational studies of natural polymers such as Cellulose and Chitin are often hindered by the practical difficulties in generating structures suitable for the simulations.
-This has motivated the recent introduction of a cellulose builder tool that generates coordinate, structure and topology files for atomistic simulations of cellulose crystal polymorphs.
-Here we present an analogous tool for Chitin, the Chitin Builder tool, a program that enhances the Visual Molecular Dynamics (VMD) environment with the ability to generate coordinate and structure files of chitin organic crystals.
-The program generates cartesian coordinates and atomic connectivity and structure files for crystalline structures of chitin polymorphs $\alpha$ and $\beta$.
-Crystal structures of any size with or without bonds in the periodic directions can be easily build.
-The resulting structure is automatically saved in PDB and PSF format (used by well known simulation packages such as NAMD) and it can be easily converted to many other file formats using VMD build-in features.
 
+Chitin is the second most abundant organic material in nature after cellulose and its study is now of great interest in the field of biocompatible and eco-friendly materials.
+Computational studies of natural polymers such as cellulose and chitin are often hindered by the practical difficulties in generating structures suitable for the simulations.
+This has motivated the recent introduction of a cellulose builder tool that generates coordinate, structure and topology files for atomistic simulations of cellulose crystal polymorphs.
+Here we present an analogous tool for chitin, the Chitin Builder tool, a program that enhances the Visual Molecular Dynamics (VMD) environment with the ability to generate coordinate and structure files of chitin organic crystals.
+The program generates Cartesian coordinates and atomic connectivity and structure files for crystalline structures of chitin polymorphs $\alpha$ and $\beta$.
+Crystal structures of any size with or without bonds in the periodic directions can be easily built.
+The resulting structure is automatically saved in Protein Data Bank (PDB) and Protein Structure Format (PSF) format (used by well-known simulation packages such as NAMD) and it can be easily converted to many other file formats using VMD build-in features.
 
 # Statement of Need
 
 Chitin is a polysaccharide present in the exoskeleton and internal structure of many invertebrates like molluscs, crustaceans, insects, fungus, algae, and other related organisms [@Dutta2002CHITINAPPLICATIONS; @Zargar2015AApplications].
 It is so prevalent in nature that it constitutes the second most abundant polymerized form of carbon on Earth.
 From the point of view of material sciences, chitin is biorenewable, environmentally friendly, biocompatible and biodegradable material.
-It has applications as a chelating agent, water treatment additive, drug carrier, biodegradable pressure‐sensitive adhesive tape, wound‐healing agents and many others [@Zargar2015AApplications; @Shamshina2019AdvancesReview; @RaviKumar2000AApplications]. 
+It has applications as a chelating agent, water treatment additive, drug carrier, biodegradable pressure‐sensitive adhesive tape, wound‐healing agents and many others [@Zargar2015AApplications; @Shamshina2019AdvancesReview; @RaviKumar2000AApplications].
 
-The possibility to generate atomic coordinates of the crystal structures of chitin is important from both a fundamental and practical point of view, since it opens the possibility to predict the properties of chitin based materials and derivatives (mechanical, thermal, interaction with solvents,...).
[…]
