### 2024-01-16 4a8644b
Minor README edits
