### 2024-01-16 soft-matter-theory-at-icmab-csic/chitin_builder#3 @amoeba
Hi @jfaraudo, thanks for fixing my usage issue and for updating the paper. I reviewed the paper and the README and sent in a small PR with edits to both: https://github.com/soft-matter-theory-at-icmab-csic/chitin_builder/pull/11.

The only remaining item from my checklist is,

> - [x] Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

From https://joss.readthedocs.io/en/latest/review_checklist.html#documentation. Can you please add a section in the README with something simple covering this?
