### 2024-01-16 3d1cb1a
Update README.md
Added a community section: instructions for contributing and support
