### 2024-01-16 2e2a605
Create CITATION.md
Added citation details
