+++ CITATION.md
@@ -1,0 +1,1 @@
+Please cite the following JOSS publication:
