### 2024-01-16 36bf535
Create GUIDELINES.md
Created community guidelines
