+++ GUIDELINES.md
@@ -1,0 +1,35 @@
+# Contributing to Chitin Builder
+
+Thank you for considering contributing to Chitin Builder! We welcome all contributions
+in the form of bug reports, feature requests, code contributions, suggestions, etc.
+
+The default branch for this repository is `main`. Please submit all pull requests
+to the `main` branch.
+
+## Bug reports and feature requests
+For bug reports and feature requests, please open an issue on the
+[GitHub issue tracker](https://github.com/soft-matter-theory-at-icmab-csic/chitin_builder/issues).
+
+If you are reporting a bug, please include the following information:
+- A short description of the bug
+- A minimal working example that reproduces the bug
+- The expected behavior
+- The actual behavior
+- The version of Chitin Builder and computer OS that you are using
+
+For feature requests, please include a short description of the feature you would
+like to see added to Chitin Builder.
+
+## Code contributions
+We welcome all code contributions in the form of pull requests. If you are
+interested in contributing code, please follow the steps below:
+- Fork the repository (default branch is `main`)
+- Create a new branch for your feature/bug fix
+- Make your changes
+- Push your changes to your fork and create a pull request
+- Assign the pull request to a reviewer and wait for the reviewers feedback
+
+## Questions and comments
+
+Please feel free to create an issue if you have any questions or comments.
+Developers can be contacted at jfaraudo@icmab.es (Jordi Faraudo)
