+++ CITATION.md
@@ -1,1 +1,3 @@
-Please cite the following JOSS publication:
+In any use of the software please cite the JOSS publication: (Submitted)
+Also, the software itself should be cited using the DOI provided by Zenodo:
+[Software DOI](https://doi.org/10.5281/zenodo.3274726)
