### 2024-01-16 b4f4a00
Update CITATION.md
