+++ CITATION.md
@@ -1,3 +1,8 @@
-In any use of the software please cite the JOSS publication: (Submitted)
+# Citing Chitin Builder
+
+In any use of the software please cite our JOSS publication: 
+David C Malaspina and Jordi Faraudo "Chitin Builder: a VMD tool for the generation of
+structures of chitin molecular crystals for atomistic simulations", JOSS (Submitted, 2023)
+
 Also, the software itself should be cited using the DOI provided by Zenodo:
 [Software DOI](https://doi.org/10.5281/zenodo.3274726)
