### 2024-01-16 ff58984
Update CITATION.md
Created a file describing how to cite this work
