+++ CITATION.md
@@ -4,5 +4,6 @@ In any use of the software please cite our JOSS publication:
 David C Malaspina and Jordi Faraudo "Chitin Builder: a VMD tool for the generation of
 structures of chitin molecular crystals for atomistic simulations", JOSS (Submitted, 2023)
 
-Also, the software itself should be cited using the DOI provided by Zenodo:
-[Software DOI](https://doi.org/10.5281/zenodo.3274726)
+In addition, the software itself should be cited using the DOI provided by Zenodo:
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3274726.svg)](https://doi.org/10.5281/zenodo.3274726)
+
