### 2024-01-16 #5771 @jfaraudo
Thanks @amoeba for your comments, with fully agree with all your points..
1) We accepted and merged your comments made to the README.m and paper.m files to the master branch.
2) We added community guidelines. A section has been added in the README.m which links to a file with a detailed description for third parties willing to contribute, report issues or seek support.
3) We have created a new software release (v1.1)

Concerning your last point ("An updated archive for JOSS") I am not sure what do you mean. If you mean the paper.md file? (this has been updated form your PR). Or maybe is something that has to be done from the journal @majensen ? Or I am missing something? (sorry, we are new to this journal)
