### 2024-01-16 de1a712
Update README.md
A minor typo has been corrected
