### 2024-01-16 6ebea4b
Update paper.md
Minor typo corrected
