### 2024-01-16 #5771 @jfaraudo
In our final check we have found a (really minor) typo in the paper, so we have corrected it. I think that now the work can be considered finished with all comments taken into account.
