### 2024-01-16 99fa5b9
Update CITATION.md
Updated JOSS link
