+++ CITATION.md
@@ -2,7 +2,7 @@
 
 In any use of the software please cite our JOSS publication: 
 David C Malaspina and Jordi Faraudo "Chitin Builder: a VMD tool for the generation of
-structures of chitin molecular crystals for atomistic simulations", JOSS (Submitted, 2023)
+structures of chitin molecular crystals for atomistic simulations", Journal of Open Source Software  [![status](https://joss.theoj.org/papers/c409d05d1e558827d1dea6275faf8965/status.svg)](https://joss.theoj.org/papers/c409d05d1e558827d1dea6275faf8965)
 
 In addition, the software itself should be cited using the DOI provided by Zenodo:
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3274726.svg)](https://doi.org/10.5281/zenodo.3274726)
