### 2024-01-16 #5771 @amoeba
Thanks @jfaraudo. My review is now an accept @majensen.

Re: the updated archive, I think that's handled automatically when the submission is accepted and I pointed it out just to make sure the new 1.1 release is used everywhere.
