### 2024-01-17 soft-matter-theory-at-icmab-csic/chitin_builder#3 @jfaraudo
@amoeba We respectfully disagree with your last comment. If you look at our README, you will see that it had a section entitled "Contributing, Reporting Issues and Support". In this section there is a short explanation and a link to an extensive document explaining in detail all these points: 1) How to contribute to the software 2) Report issues or problems with the software 3) Seek support. This is fully compliant with the journal rules  and in fact it is identical to what is done exactly in other JOSS publications (see for example this recently published paper in JOSS https://github.com/cmelab/flowerMD ). So we think that the remaining item in the checklist should be closed, please.
