### 2024-01-17 soft-matter-theory-at-icmab-csic/chitin_builder#3 @amoeba
Hi @jfaraudo. My comment was made before the section was added in https://github.com/soft-matter-theory-at-icmab-csic/chitin_builder/commit/3d1cb1a250b03eeeeed28798a2a72754db07acbc but now that it has been added I've checked the item above.

My review is an Accept at this stage, see https://github.com/openjournals/joss-reviews/issues/5771#issuecomment-1894449823, and this issue can be closed if you like.
