### 2024-01-19 #5771 @majensen
@jfaraudo Thanks for all the work. I will do a quick editorial runthrough of the paper itself. That may generate a couple of fixes needed. Re archive, what JOSS requires is an open archive of your repository, in [Zenodo](https://zenodo.org) or similar service. Have a look at [this tutorial](https://github.com/openjournals/joss-reviews/issues/1839#issuecomment-560048729) from another JOSS review.
