### 2024-01-21 #5771 @majensen
Thanks @jfaraudo  - the journal prefers that the title and authors on the archive match those on the paper, if you can update those items.
