### 2024-01-22 #5771 @jfaraudo
OK @majensen The zenodo archive was generated automatically after creating the new software version (we had an automatic update of our GitHub repo in Zenodo) and it takes as title the GitHub repo name and as authos everyone that contributed in the repo (including referees like @amoeba that did a pull request on GitHub). We will try to correct this today. If a correction is not possible we will create manually another Zenodo archive for the JOSS publication.
