### 2024-01-22 #5771 @jfaraudo
@majensen Corrected, now authors and title matches that of JOSS submission and files correspond to latest repository version . Same DOI:  https://zenodo.org/doi/10.5281/zenodo.3274725
