### 2024-01-27 #5771 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
