### 2024-01-29 #5771 @Kevin-Mattheus-Moerman
@jfaraudo  As AEiC I will now help to process final steps towards acceptance in JOSS. I have just checked your repository, this review, the archive link, and the paper. All seems in order, so I will now proceed to accept this work.
